Refactored `encoder` module into v1_0 and v1_1 submodules
